# Update manager refactor