--- a/UnityProject/Assets/Scripts/Managers/UpdateManager/UpdateManager.cs.meta
+++ b/UnityProject/Assets/Scripts/Managers/UpdateManager/UpdateManager.cs.meta
@@ -1,9 +1,10 @@
 fileFormatVersion: 2
 guid: 5becb0357443a41449a1dc06b7b68eae
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
-  executionOrder: 0
+  executionOrder: -50
   icon: {instanceID: 0}
   userData: 
   assetBundleName: