--- a/UnityProject/Assets/Scripts/UI/PlayerHealthUI.cs
+++ b/UnityProject/Assets/Scripts/UI/PlayerHealthUI.cs
@@ -27,15 +27,12 @@ public class PlayerHealthUI : MonoBehaviour
 
 	private void OnEnable()
 	{
-		UpdateManager.Instance.Add(UpdateMe);
+		UpdateManager.Add(CallbackType.UPDATE, UpdateMe);
 	}
 
 	private void OnDisable()
 	{
-		if (UpdateManager.Instance != null)
-		{
-			UpdateManager.Instance.Remove(UpdateMe);
-		}
+		UpdateManager.Remove(CallbackType.UPDATE, UpdateMe);
 	}
 
 	private void DisableAll()