--- a/UnityProject/Assets/Scripts/UI/UI_OxygenAlert.cs
+++ b/UnityProject/Assets/Scripts/UI/UI_OxygenAlert.cs
@@ -23,15 +23,12 @@ public class UI_OxygenAlert : TooltipMonoBehaviour
 
 	private void OnEnable()
 	{
-		UpdateManager.Instance.Add(UpdateMe);
+		UpdateManager.Add(CallbackType.UPDATE, UpdateMe);
 	}
 
 	private void OnDisable()
 	{
-		if (UpdateManager.Instance != null)
-		{
-			UpdateManager.Instance.Remove(UpdateMe);
-		}
+		UpdateManager.Remove(CallbackType.UPDATE, UpdateMe);
 		ResetImg();
 	}